requirements.txt
torch>=1.10.0
torchvision>=0.11.0
ultralytics>=8.0.0
streamlit>=1.28.0
Pillow>=8.3.0
opencv-python>=4.5.0
numpy>=1.21.0
pandas>=1.3.0
